requirements.txt
pandas==2.1.2
python-dotenv==1.0.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.9
sqlalchemy.orm==1.2.10
